requirements.txt
iBeaker==1.7.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psutil==3.1.1
pytz==2015.4
redis==2.10.3
Werkzeug==0.10.4
wheel==0.24.0
